refactor(core): move tasks/ from under agents/ to top-level
The new task abstraction holds shells, monitors, and dreams alongside
agents. Living under agents/ misnamed the boundary — three of the
four kinds aren't agents at all. Move to packages/core/src/tasks/
to match what the abstraction actually represents.
