+++ packages/core/src/agents/index.ts
@@ -17,10 +17,3 @@ export * from './backends/index.js';
 export * from './arena/index.js';
 export * from './runtime/index.js';
 export * from './background-agent-resume.js';
-export * from './tasks/types.js';
-export * from './tasks/agent-task.js';
-export * from './tasks/shell-task.js';
-export * from './tasks/monitor-task.js';
-export * from './tasks/dream-task.js';
-export * from './tasks/registry.js';
-export * from './tasks/dispatcher.js';
