+++ packages/core/src/tasks/agent-task.ts
@@ -18,10 +18,10 @@
  * grace timer; the registry is kind-agnostic.
  */
 
-import { createDebugLogger } from '../../utils/debugLogger.js';
-import { escapeXml } from '../../utils/xml.js';
-import { patchAgentMeta } from '../agent-transcript.js';
-import type { AgentExternalInput } from '../runtime/agent-types.js';
+import { createDebugLogger } from '../utils/debugLogger.js';
+import { escapeXml } from '../utils/xml.js';
+import { patchAgentMeta } from '../agents/agent-transcript.js';
+import type { AgentExternalInput } from '../agents/runtime/agent-types.js';
 import type { TaskBase, TaskRegistration, TaskStatus } from './types.js';
 import type { TaskRegistry } from './registry.js';
 import type { Task } from './dispatcher.js';
