+++ packages/core/src/tasks/shell-task.test.ts
